Fix all the Linux warnings. This introduces some new warnings under OS X, and Windows still has unfixed warnings. It will obviously take some iterations to get them all.

+++ org/antlr/v4/runtime/ANTLRInputStream.cpp
@@ -88,7 +88,7 @@ namespace org {
                            int numRead = 0;
                            int p = 0;
                            do {
-                               if (p + readChunkSize > data.length()) { // overflow?
+			      if (p + readChunkSize > (int)data.length()) { // overflow?
                                    // System.out.println("### overflow p="+p+", data.length="+data.length);
                                    data = antlrcpp::Arrays::copyOf(data, (int)data.length() * 2);
                                }
